(Dispatch Org)
reorganizing mqtt and dispatch functionalities.

- merged `mqtt.py` and `dispatch.py` into single `dispatch.py` module.
- Deleted `mqtt.py` standalone module
- incorporated dispatcher + messaging into main method.
